[Profiler][Minor] Organize collection.h/.cpp
Pull Request resolved: https://github.com/pytorch/pytorch/pull/82992

Collection of Torch ops is quite complex compared to backend events / allocations / ooms. Python is also complex, however it is already factored into a standalone unit. This PR just shuffles the contents of collection.cpp to group the Torch op specific parts together, and does various cleanups to the code.
ghstack-source-id: 165143014

Differential Revision: [D38426344](https://our.internmc.facebook.com/intern/diff/D38426344/)