[wip] remove DBR quantization from the codebase
Summary:

DBR quantization is a no-go for now because it does not align well with
PyTorch 2.0 plans and we do not want to build yet another tracing system.

Deleting it from the codebase for now since there are no plans to develop
this in the near future. We can bring it back at a later time if necessary.

Test plan:

CI

ghstack-source-id: 1c5296e857d2a6ca2ad499a9f3956d816d9d9c89
Pull Request resolved: https://github.com/pytorch/pytorch/pull/83642